move test files into examples test dir
